free all

+++ parssing/check_quotes.c
@@ -1,8 +1,8 @@
 #include "../includes/minishell.h"
 
-int check_quotes(char *cmd, int i)
+int check_quotes(char *cmd, size_t i)
 {
-  int j;
+  size_t j;
   int flag;
 
   j = 0;
